// src-tauri/bindings/export/ColorTheme.d.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BookmarkItemColorTheme } from "../BookmarkItemColorTheme";
import type { PredicateInputBox } from "../PredicateInputBox";
import type { SuggestionWindow } from "../SuggestionWindow";
import type { TagItem } from "../TagItem";

export type ColorTheme = { bookmarkItem: BookmarkItemColorTheme, suggestionWindow: SuggestionWindow, predicateInputBox: PredicateInputBox, tagItem: TagItem, bg: string, };